# Implemented serializers caching for lookup