--- a/core/commonMain/src/kotlinx/serialization/Serializers.kt
+++ b/core/commonMain/src/kotlinx/serialization/Serializers.kt
@@ -209,6 +209,7 @@ private fun KClass<Any>.compiledParametrizedSerializer(serializers: List<KSerial
     return constructSerializerForGivenTypeArgs(*serializers.toTypedArray())
 }
 
+@OptIn(ExperimentalSerializationApi::class)
 private fun KClass<Any>.builtinParametrizedSerializer(
     typeArguments: List<KType>,
     serializers: List<KSerializer<Any?>>,