… benchmark/src/jmh/kotlin/kotlinx/benchmarks/json/LookupOverheadBenchmark.kt
… core/commonMain/src/kotlinx/serialization/internal/Platform.common.kt
… core/jsMain/src/kotlinx/serialization/internal/Platform.kt
… core/jvmMain/src/kotlinx/serialization/internal/Caching.kt
… core/jvmMain/src/kotlinx/serialization/internal/Platform.kt
… core/nativeMain/src/kotlinx/serialization/internal/Platform.kt