@qwwdfsad 2022-08-18
Results:

```
Baseline
Benchmark                                Mode  Cnt   Score   Error   Units
LookupOverheadBenchmark.dataPlain       thrpt   14  20.943 ± 0.077  ops/us
LookupOverheadBenchmark.dataReified     thrpt   14   4.777 ± 0.028  ops/us
LookupOverheadBenchmark.genericPlain    thrpt   14   8.740 ± 0.022  ops/us
LookupOverheadBenchmark.genericReified  thrpt   14   0.323 ± 0.014  ops/us
LookupOverheadBenchmark.objectPlain     thrpt   14  48.704 ± 0.073  ops/us
LookupOverheadBenchmark.objectReified   thrpt   14   0.927 ± 0.004  ops/us

This patch
Benchmark                                Mode  Cnt   Score   Error   Units
LookupOverheadBenchmark.dataPlain       thrpt   14  21.232 ± 0.057  ops/us
LookupOverheadBenchmark.dataReified     thrpt   14   6.002 ± 0.126  ops/us
LookupOverheadBenchmark.genericPlain    thrpt   14   8.466 ± 0.180  ops/us
LookupOverheadBenchmark.genericReified  thrpt   14   0.589 ± 0.005  ops/us
LookupOverheadBenchmark.objectPlain     thrpt   14  48.713 ± 0.059  ops/us
LookupOverheadBenchmark.objectReified   thrpt   14   7.837 ± 0.229  ops/us

This patch + new kotlin.reflect
Benchmark                                Mode  Cnt   Score   Error   Units
LookupOverheadBenchmark.dataPlain       thrpt   14  21.241 ± 0.245  ops/us
LookupOverheadBenchmark.dataReified     thrpt   14  15.708 ± 0.210  ops/us
LookupOverheadBenchmark.genericPlain    thrpt   14   8.597 ± 0.145  ops/us
LookupOverheadBenchmark.genericReified  thrpt   14   2.143 ± 0.002  ops/us
LookupOverheadBenchmark.objectPlain     thrpt   14  48.793 ± 0.242  ops/us
LookupOverheadBenchmark.objectReified   thrpt   14  30.176 ± 0.219  ops/us
```