@shanshin 2022-08-25
Results with latest changes 
```
Benchmark                                Mode  Cnt   Score   Error   Units
LookupOverheadBenchmark.dataPlain       thrpt   14  21.152 ± 0.495  ops/us
LookupOverheadBenchmark.dataReified     thrpt   14  16.309 ± 0.044  ops/us
LookupOverheadBenchmark.genericPlain    thrpt   14   8.232 ± 0.079  ops/us
LookupOverheadBenchmark.genericReified  thrpt   14   9.537 ± 0.079  ops/us
LookupOverheadBenchmark.objectPlain     thrpt   14  48.672 ± 0.086  ops/us
LookupOverheadBenchmark.objectReified   thrpt   14  28.955 ± 1.601  ops/us
```